src/requirements.txt
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Paginate==0.5.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
GeoAlchemy2==0.5.0
Shapely==1.6.4.post2
WTForms==2.2.1
astroplan==0.6
astropy==4.0
boto3==1.9.11
botocore==1.12.11
colorlog==4.0.2
fastavro==0.21.8
kafka==1.3.5
kafka-python==1.4.3
lcogt-logging==0.3.2
lz4tools==1.3.1.2
matplotlib==3.1.2
psycopg2-binary==2.8.4
pyjwt==1.7.1
pytz==2018.9
requests==2.22.0
sqlalchemy==1.3.12
urllib3==1.21
werkzeug==0.16.0
